+++ src/Auth/index.js
@@ -1,23 +1,8 @@
-import AdminOnly from './AdminOnly';
 import APIEntity from './APIEntity';
 import ObjectAsList from './ObjectAsList';
 import DebugObject from './DebugObject';
-import EmailVerifiedOnly from './EmailVerifiedOnly';
 import LoadingCover from './LoadingCover';
-import LoggedInOnly from './LoggedInOnly';
-import LoggedOutOnly from './LoggedOutOnly';
 import Status from './Status';
 import User from './User';
 
-export {
-  APIEntity,
-  AdminOnly,
-  DebugObject,
-  LoadingCover,
-  EmailVerifiedOnly,
-  LoggedInOnly,
-  LoggedOutOnly,
-  ObjectAsList,
-  Status,
-  User,
-};
+export { APIEntity, DebugObject, LoadingCover, ObjectAsList, Status, User };