+++ src/Footer/stories.js
@@ -1,8 +1,0 @@
-/* @flow strict */
-
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-
-import { Footer } from './index';
-
-storiesOf('Footer', module).add('default', () => <Footer />);